requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
